pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hanimo-rag"
version = "0.1.0"
description = "PostgreSQL-native hybrid RAG engine with vector + full-text + knowledge graph search"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "hanimo Team"}
]
keywords = ["rag", "retrieval", "augmented", "generation", "pgvector", "postgresql", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "asyncpg>=0.29.0",
    "pypdf>=4.0.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.23",
    "httpx>=0.27.0",
    "pydantic-settings>=2.2.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
local = [
    "sentence-transformers>=3.0.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx",
    "ruff>=0.3.0",
]

[project.scripts]
hanimo-rag = "hanimo_rag.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hanimo_rag*"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
